+++ src/main/java/com/ndashimye/firstapp/usertodo/UserTodoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserTodoRepository extends JpaRepository<UserTodo, Integer> {
+public interface UserTodoRepository extends JpaRepository<UserTodo, Long> {
 
 }
